writes test to change the average_goals_against from a stat_tracker method to a team method
